server/requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
pydantic
langchain
boto3
langchain-community
python-dotenv
openai